logger creation moved into utils/ and used at net_eye.py
