+++ net_eye.py
@@ -1,20 +1,5 @@
 #! /usr/bin/env python3
-import logging
-
-
-def create_logger(name):
-    '''
-
-    '''
-    logger = logging.getLogger(name)
-    logger.setLevel(logging.DEBUG)
-    ch = logging.StreamHandler()
-    ch.setLevel(logging.DEBUG)
-    formatter = logging.Formatter(
-            '%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    ch.setFormatter(formatter)
-    logger.addHandler(ch)
-    return logger
+from utils.net_logger import create_logger
 
 
 def main(argv):
